+++ CHANGELOG.md
@@ -9,10 +9,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ---
 
-## [0.10.0] - (06-02-2021)
+## [0.10.0] - (06-10-2021)
 
 - added journal length to logging output
-  - this is so i can see how the singleton instance is workign across projects using the logger
+  - this is to see how the singleton instance is working across projects using the logger
 - updated tests to utilize new logger
 - added test rest server
   - this is the beginning of something that the multipart download tests can run against
@@ -22,5 +22,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - added DO rpm v1.19.0-2 to testing artifacts since it is a couple of Mb (not too small, not too big)
 - looking into building packages (npm pack) to upload as releases
   - these releases will produces "tags" for the version and allow for private versioning without publishing to NPM
-- finished DO class
+- finished DO class methods/functions
 - Moved more as3 models/functions from vscode-f5 (re-imported/re-factored)
+- configured TEEM agent inject to as3/do declartion post
+- added schema and examples endpoints to atcMetaData constants
+- updated f5Client.discover to output DiscoverInfo